requirements.txt
# fmt: off
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.1
pydantic==2.9.0
sqlalchemy==2.0.35
python-jose==3.3.0y
passlib==1.7.4
asyncpg==0.30.0
psycopg==3.2.3
psycopg-pool==3.2.4
psycopg2-binary==2.9.10
newsapi-python==0.2.7
beautifulsoup4==4.12.3
scalar_fastapi==1.0.3
python-multipart==0.0.19
sentry-sdk==2.22.0
bcrypt==4.0.1